Add arbitrary (sync, spawn) command execution
This commit adds two additional keywords (`exec:`, `spawn:`) for
executing arbitrary commands when defining key maps.

There are three special variables that are expanded into actual values.

- `%vault`: The name of the vault
- `%note`: The name of the selected note
- `%note_path`: The absolute path to the selected note

For example the command can be used to open obsidian using the obsidian
uri pattern:

```
  { key = "ctrl+alt+o", command = "spawn:open obsidian://open?vault=%vault&file=%note" },
```

Or external editor like `nvim`:

```
  { key = "ctrl+alt+e", command = "exec:nvim %note_path" },
```

Changelog: added
Package: basalt
